refactor: clean up backend AM-3258

+++ gravitee-am-repository/gravitee-am-repository-jdbc-api/src/main/java/io/gravitee/am/repository/jdbc/provider/utils/ObjectUtils.java
@@ -46,7 +46,7 @@ public final class ObjectUtils {
                     } else {
                         Long myLong = Long.valueOf(string);
                         if (string.equals(myLong.toString())) {
-                            if (myLong == (long)myLong.intValue()) {
+                            if (myLong == myLong.intValue()) {
                                 return myLong.intValue();
                             }
